Move Tethys files to dedicated directories
