+++ lib/CMakeLists.txt
@@ -16,8 +16,8 @@ add_library(
     janus/janus_client.cpp
     janus/janus_server.cpp
     oceanus/cuckoo.cpp
-    oceanus/tethys_graph.cpp
-    oceanus/tethys_allocator.cpp
+    tethys/tethys_graph.cpp
+    tethys/tethys_allocator.cpp
 )
 # Add an alias
 add_library(OpenSSE::schemes ALIAS schemes)
