+++ test/tethys_graph.cpp
@@ -1,4 +1,4 @@
-#include <sse/schemes/oceanus/details/tethys_graph.hpp>
+#include <sse/schemes/tethys/details/tethys_graph.hpp>
 
 #include <iostream>
